requirements.txt
torch==1.11.0
torchvision==0.12.0
pypng==0.0.21